Don't delete email tmpfile unless sendmail returns 0
I was hitting the following in my logs and needed to analyze:

No recipient addresses found in header
OK

+++ borgsphere.sh
@@ -38,7 +38,7 @@ send_mail()
       /bin/echo -e "$4\r\n" >> "$tmpfile"
    fi
    /usr/sbin/sendmail -t < "$tmpfile"
-   rm $tmpfile
+   [ $? -eq 0 ] && rm $tmpfile
    echo "OK"
 }
